Integrate glibc Fixes, Remove Obsolete Code
  * Makefile.in
    - Add useful compile flags

  * conv.c
    - Delete duplicate nMatch case in Bconv()

  * input.c
    - Notate #endif statements
    - Delete SGI IRIX block
    - Noted that esgetenv() may be able to be safely removed

  * prim-etc.c
    - Pull in changes by Alex Thola from glibc-fixes git branch

  * prim-sys.c
    - Pull in changes by Alex Thola from glibc-fixes git branch
    - Change always false mode_t comparison to check against
      `(mode_t)(~0)` instead of `-1`

  * print.c
    - `FMT_short` in `initconv()` now uses an `int` type instead of `short`,
      this should preserve expected behavior without relying on UB

  * proc.c
    - Pull in changes by Alex Thola from glibc-fixes git branch

  * status.c
    - Pull in changes by Alex Thola from glibc-fixes git branch

  * stdenv.h
    - Add a macro to test for GCC use and disable GNU C extensions
    - Added test for C++ compiler and forced throwing an error, as the
      codebase is modernized, the potential benefits of compilation as C++
      will diminish rapidly
    - Remove conditional compilation for SGI IRIX and HP/UX
    - Made the inclusion of `<sys/wait.h>` only occur on non-linux
      platforms

  * version.c
    - Minor bump in version string

+++ print.c
@@ -90,7 +90,7 @@ static void intconv(Format *format, unsigned int radix, int upper, char *altform
 	if (flags & FMT_long)
 		n = va_arg(format->args, long);
 	else if (flags & FMT_short)
-		n = va_arg(format->args, short);
+		n = va_arg(format->args, int);
 	else
 		n = va_arg(format->args, int);
